Rename ext/elasticsearch to opensearch
Rename ext/elasticsearch to ext/opensearch and replace code imports and references to ext.elasticsearch.
