+++ hepdata/ext/opensearch/config/record_mapping.py
